ajustes migrations e frontend

+++ backend/migrations/20210312145107_create_doacoes_com_metas_objetivos_table.js
@@ -4,7 +4,7 @@ exports.up = function(knex) {
         table.increments('id').primary()
         table.string('titulo').notNull()
         table.string('descricao', 250)
-        table.decimal('valor').notNull()
+        table.integer('valor').notNull()
         table.integer('metasId').references('id')
         .inTable('doacoes_com_metas').notNull()
         table.integer('userId').references('id')
